requirements.txt
absl-py~=2.1.0
numpy~=2.1.0
scipy~=1.14.1
highs~=1.7.2
highspy~=1.7.2
osqp~=0.6.3
qdldl~=0.1.7
qpsolvers~=4.3.3
tzdata~=2024.1
six~=1.16.0
dateutil~=2.9.0
python-dateutil>=2.8.2
pytz~=2024.1
et-xmlfile~=1.1.0
openpyxl~=3.1.5
pandas~=2.2.2
pytest~=8.3.2
lxml~=5.3.0
googleapis-common-protos~=1.65.0
pyasn1~=0.6.1
cachetools~=6.0.0
certifi~=2025.4.26
charset-normalizer~=3.4.2
requests>=2.32.4
google-api-core~=2.24.2
google_auth~=2.40.2
google-cloud-core~=2.4.3
google-cloud-storage~=3.1.0
google-crc32c~=1.7.1
google_resumable_media~=2.7.2
grpcio==1.72.2
idna~=3.10
proto-plus~=1.26.1
pyasn1-modules~=0.4.2
rsa~=4.9.1
urllib3>=2.5.0
# Must be compatible with the protobuf version declared with bazel_dep in MODULE.bazel
# See https://github.com/protocolbuffers/protobuf/blob/v29.3/python/google/protobuf/__init__.py#L10
protobuf>=5.29.5
black>=24.3.0
